# goios-peer/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/device/{udid}/apps": {
            "post": {
                "description": "List the installed apps on a device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "List apps on a device",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/installationproxy.AppInfo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    }
                }
            }
        },
        "/device/{udid}/apps/install": {
            "post": {
                "description": "Install app on a device by uploading an ipa file",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "Install app on a device",
                "parameters": [
                    {
                        "type": "file",
                        "description": "ipa file to install",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    }
                }
            }
        },
        "/device/{udid}/apps/kill": {
            "post": {
                "description": "Kill running app on a device by provided bundleID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "Kill running app on a device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "bundle identifier of the targeted app",
                        "name": "bundleID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    }
                }
            }
        },
        "/device/{udid}/apps/launch": {
            "post": {
                "description": "Launch app on a device by provided bundleID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "Launch app on a device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "bundle identifier of the targeted app",
                        "name": "bundleID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.GenericResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "installationproxy.AppInfo": {
            "type": "object",
            "additionalProperties": {}
        },
...
    }
}